Show the image of the page even when the folder is private.
Previously, from a public default page in a private folder, we tried
to traverse to the full path of `@@images`, which meant we needed to
traverse to the private folder, which failed, so we showed the
fallback image.  Instead, we now simply traverse to `@@images`
directly, which means we can show the image of the page.